LayoutTests/http/tests/security/mixedContent/insecure-image-in-iframe.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
}
</script>
<p>This test loads a secure iframe that loads an insecure image.  We should
*not* get a mixed content callback becase the main frame is HTTP and the image
doesn't contaminate the child iframe's security origin with mixed content.</p>
<iframe src="https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html";
></iframe>
</body>
</html>
